[ias-verify] fix potential index out of bound error in `safe_indexing`
